Split out dtest's k3s functionality into a spearate dtest_k3s package to avoid circular imports when using from kates. Also use dexec instead of supervisor and turn off logging by default to make output less noisy.
