+++ pkg/dtest_k3s/k3s_test.go
@@ -1,4 +1,4 @@
-package dtest
+package dtest_k3s
 
 import (
 	"os"
